Refactor and optimize attendance tracking system
- Streamline imports across multiple files to remove unused imports
- Update AttendanceController and ClassroomController for better code organization
- Modify TeacherController to use Long instead of Integer for ID parameters
- Refactor UploadController to handle single file uploads instead of multiple files
- Update FileUploadInput to use a single FileUpload instead of a List
- Remove unused imports and interfaces from various DTO and entity classes
- Optimize AttendanceServiceImpl:
  - Improve error handling and input validation
  - Enhance performance of database queries
  - Refactor methods for better readability and maintainability
- Update TeacherServiceImpl to use Optional for better null handling
- Refactor AttendanceWebSocket and RealTimeAttendanceWebSocket for improved WebSocket handling
- Remove unused YAML and SQL files
- Update .gitignore to exclude uploads directory and backup files

This commit focuses on code optimization, error handling improvements, and overall system refactoring to enhance performance and maintainability of the attendance tracking system.

+++ src/main/java/com/pshs/ams/controllers/TeacherController.java
@@ -36,7 +36,7 @@ public class TeacherController {
 
 	@GET
 	@Path("/{id}")
-	public Response getTeacherById(@PathParam("id") Integer id) {
+	public Response getTeacherById(@PathParam("id") Long id) {
 		if (id <= 0) {
 			return Response.status(Response.Status.BAD_REQUEST).entity(new MessageDTO("Invalid id", CodeStatus.BAD_REQUEST)).build();
 		}
@@ -96,7 +96,7 @@ public class TeacherController {
 
 	@PUT
 	@Path("/update/{id}")
-	public Response updateTeacher(TeacherDTO teacherDTO, @PathParam("id") Integer id) {
+	public Response updateTeacher(TeacherDTO teacherDTO, @PathParam("id") Long id) {
 		if (teacherDTO == null || id <= 0) {
 			return Response.status(Response.Status.BAD_REQUEST).entity(new MessageDTO("Teacher cannot be null or id cannot be less than or equal to zero", CodeStatus.BAD_REQUEST)).build();
 		}
